Remove dependencies provided by well-vendored package
We have a dependency on golang-github-cpuguy83-go-md2man,
which itself depends on golang-github-russross-blackfriday and
golang-github-shurcooL-sanitized_anchor_name, so we included all three
as BuildRequires.  As of the 1.0.6 release of
golang-github-cpuguy83-go-md2man though, the other two are vendored
inside golang-github-cpuguy83-go-md2man.  This means we can drop the
other two dependencies.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>