Remove cc-runtime detection
Instead of patching docker to handle defaulting to use the Clear
Containers runtime if it was available and /dev/kvm existed, move
detection and default behavior into the kata-runtime package. That
change uses systemd unit configuration and is less invasive and
doesn't require as much maintenance as keeping up this patch set
did. It also handles setting overlay2 for the storage driver.